feat(tips): rework the tip dialog into a pill-anchored card (owner pick C) (#361)
The A/B/C footer mock went to the owner on the agent canvas; the pick was C.
TipModal -> TipCard: no overlay, no aria-modal, the app stays live behind it.
Anchored to the dock pill in both sidebar variants (ResizeObserver re-anchors
across collapse/expand; bottom-left fallback), pill click toggles, the ... menu
moved to the header, the footer stays at three nowrap buttons, and Next tip
advances in place, closing the card when the rotation runs dry.
